usb_canbus: Send echo frame before processing the frame
The Linux kernel support reports a canbus message as transmitted when
it gets the echo frame back.  Processing the message prior to sending
the echo frame can lead to odd looking debugging logs (as the response
messages may appear to predate the request messages).  This doesn't
impact the Klipper code, but it does make analyzing logs harder.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>